+++ src/components/Main.js
@@ -5,6 +5,7 @@ import Videos from './Videos';
 import Login from './Login/Login';
 import Contact from './Contact';
 import Task from './Task'
+import Profile from './Profile'
 
 const Main = () => (
   <Switch>
@@ -13,6 +14,7 @@ const Main = () => (
     <Route path="/login" exact component={Login} />
     <Route path="/contact" exact component={Contact} />
     <Route path='/task' component={Task} />
+    <Route path='/profile' component={Profile} />
   </Switch>
 )
